requirements.txt
python-telegram-bot==22.7
python-dotenv==1.2.2
requests==2.32.5
sentence-transformers==5.3.0
numpy==2.4.3
